requirements.txt
Django>=3.2.13,<3.3.0
djangorestframework>=3.13.1,<3.14.0

psycopg2>=2.9.3,<2.10.0

pandas>=1.4.2,<1.5.0
openpyxl>=3.0.9,<3.1.0

flake8>=4.0.1,<4.1.0